Extract region from project ID
