+++ descopesdk/src/test/java/com/descope/internal/others/JsonTest.kt
@@ -1,6 +1,5 @@
-package com.descope.session
+package com.descope.internal.others
 
-import com.descope.internal.others.toMap
 import org.json.JSONObject
 import org.junit.Assert.*
 import org.junit.Test
